Axum

+++ src/database.rs
@@ -1,5 +1,4 @@
 use mongodb::{bson::doc, options::ClientOptions, Collection};
-use rocket::futures::StreamExt;
 
 use crate::models::Ban;
 
@@ -21,15 +20,15 @@ impl Database {
 
         Self { bans: col }
     }
-    pub async fn get_all_bans(&self) -> Vec<Ban> {
-        self.bans
-            .find(None, None)
-            .await
-            .unwrap()
-            .map(|doc| doc.unwrap())
-            .collect()
-            .await
-    }
+    // pub async fn get_all_bans(&self) -> Vec<Ban> {
+    //     self.bans
+    //         .find(None, None)
+    //         .await
+    //         .unwrap()
+    //         .map(|doc| doc.unwrap())
+    //         .collect()
+    //         .await
+    // }
 
     pub async fn find_active_ban(&self, user_id: u64) -> Option<Ban> {
         let now = chrono::Utc::now();
@@ -53,9 +52,14 @@ impl Database {
         }
     }
 
-    pub async fn remove_ban(&self, user_id: u64) {
-        let filter = doc! { "userId": user_id as i64 };
-        self.bans.delete_one(filter, None).await.unwrap();
+    pub async fn remove_ban(&self, user_id: u64) -> Result<(), String> {
+        if self.find_active_ban(user_id).await.is_none() {
+            Err("User is not banned".to_string())
+        } else {
+            let filter = doc! { "userId": user_id as i64 };
+            self.bans.delete_one(filter, None).await.unwrap();
+            Ok(())
+        }
     }
 }
